requirements.txt
boto3==1.5.6
botocore==1.8.20
Django==2.0
django-bootstrap4==0.0.4
django-disqus==0.5
django-extensions==1.9.8
django-filter==1.1.0
django-scss==0.6.1
django-storages==1.6.5
django-taggit==0.22.1
djangorestframework==3.7.3
docutils==0.14
jmespath==0.9.3
mock==2.0.0
numpy==1.13.3
olefile==0.44
pbr==3.1.1
Pillow==4.3.0
psycopg2==2.7.3.2
python-dateutil==2.6.1
pytz==2017.3
s3transfer==0.1.12
six==1.11.0
typing==3.6.2
